microsoft-fabric/integracio_per_consultar_dades/exemples/jdbc/requirements.txt
azure-identity
jaydebeapi
JPype1
msal 
requests
pandas